2 rooms
Two rooms were created. In the first room, the player must use a portal to collect the balls and defeat the enemy. Once accomplished, the door will open. In the second room, there is a water pool that gradually decreases the player's health. To proceed to room 3, the player must utilize portals.